UXDD_GCcollab_Request_Membership_v01.pdf
UX design document has been created as a reference point and guideline to assist with the development of the new Request a Membership feature for GCcollab. This is a pilot version of a possible end to end UX process meant to guide the design of new and existing features, apps and microservices.

The current version is a PDF which hopefully can have comments and notes added to it by team members to enhance and strengthen the process while helping improve task clarity.